requirements.txt
click==8.0
cssselect==1.1.0
Flask==2.2.2
itsdangerous==2.0
Jinja2==3.0
lxml==4.9.2
MarkupSafe==2.1.1
parsel==1.5.2
python-dateutil==2.8.1
PyYAML==5.3.1
selectorlib==0.16.0
six==1.14.0
w3lib==1.21.0
Werkzeug==2.2.2
requests==2.28.2
openai==0.26.5
spacy==3.5.0
numpy==1.24.2

-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchtext==0.14.1
matplotlib==3.7.1